% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_deps.R
\name{getDirectDeps}
\alias{getDirectDeps}
\title{Installation data for \pkg{patRoon} and its direct dependencies}
\usage{
getDirectDeps()
}
\value{
A \code{list} with miscellaneous package information.
}
\description{
Returns information for \pkg{patRoon} and its direct \R package dependencies.
}
\details{
The information returned by this function may change over time, and is downloaded from the internet.
}
