% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installing}
\alias{installing}
\alias{install}
\alias{update}
\alias{sync}
\title{Managing \pkg{patRoon} installations}
\usage{
install(
  origin = NULL,
  pkgs = NULL,
  ignorePkgs = NULL,
  lib.loc = NULL,
  allDeps = FALSE,
  ask = TRUE,
  force = FALSE,
  quiet = TRUE
)

update(
  origin = NULL,
  pkgs = NULL,
  ignorePkgs = NULL,
  lib.loc = NULL,
  allDeps = FALSE,
  ask = TRUE,
  quiet = TRUE
)

sync(
  origin = NULL,
  pkgs = NULL,
  ignorePkgs = NULL,
  lib.loc = NULL,
  allDeps = FALSE,
  ask = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{origin}{Where \pkg{patRoon} and its dependencies are installed from. Valid values are: \code{"patRoonDeps"},
\code{"runiverse"}, \code{"regular"}. If \code{NULL} then the default on Windows is \code{"patRoonDeps"} and \code{"runiverse"} otherwise.
See below for more details.}

\item{pkgs}{A \code{character} vector with a subset of packages to process. If \code{NULL} then all packages are considered.}

\item{ignorePkgs}{A \code{character} vector with packages that will not be considered. Can also be \code{"big"} to exclude
large packages (\emph{e.g.} \pkg{patRoonDeps}). If \code{NULL} then no packages will be ignored.}

\item{lib.loc}{The path to the \R library where packages will be installed. Set to \code{NULL} for the default \R library.}

\item{allDeps}{Consider \emph{all} dependencies, including recursive dependencies. This is mainly useful for \code{sync()}.
Currently only supported for \code{origin="patRoonDeps"} and when \code{pkgs}/\code{ignorePkgs} are both \code{NULL}.}

\item{ask}{Set to \code{TRUE} to ask before proceeding package installations. No effect on non-interactive \R sessions.}

\item{force}{If \code{TRUE} then packages will \emph{always} be installed, even if already present and with the correct
version.}

\item{quiet}{If \code{TRUE} the installations are performed more quietly (sets the \code{quiet} option to
\code{\link[=install.packages]{install.packages()}}).}
}
\description{
Functions to install, update and synchronize \pkg{patRoon} and its dependencies from various repositories.
}
\details{
Details here:
\itemize{
\item Explain functions
\item Explain origins
}
}
