% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{legacy}
\alias{legacy}
\alias{toggleLegacy}
\alias{removeLegacy}
\title{Manage legacy \pkg{patRoon} installations}
\usage{
toggleLegacy(enable = NULL, ask = TRUE)

removeLegacy(restoreRProfile = FALSE, ask = TRUE)
}
\arguments{
\item{enable}{Enables (\code{enable=TRUE}), disables (\code{enable=FALSE}) or toggles (\code{enable=NULL}) a legacy \pkg{patRoon}
installation.}

\item{ask}{Set to \code{TRUE} to ask before proceeding. No effect on non-interactive \R sessions.}

\item{restoreRProfile}{If \code{TRUE} then the modifications to the user's \code{.Rprofile} file will be removed that were
automatically performed by the legacy installation script.}
}
\value{
All functions return \code{NULL} invisibly.
}
\description{
Toggles or removes installations of \href{https://github.com/rickhelmus/patRoon}{patRoon} that were installed by the
legacy installation script (\code{install_patRoon.R}).
}
\details{
Previous versions of \pkg{patRoon} (\verb{<2.3}) could be installed \emph{via} a (now legacy) installation \R script
(\code{install_patRoon.R}). The functions documented here can (temporarily) undo the changes by this script, so that the
\link[=installing]{installation functions in this package} can be used to manage \pkg{patRoon} installations.

The \code{toggleLegacy()} function is used to enable or disable legacy installations temporarily.

The \code{removeLegacy()} function is used to undo legacy installations.
}
\note{
The \R session should be restarted to make the changes effective.
}
\section{Background}{
 This section is purely informative, but may be of use when you want to manually manage legacy
installations.

The legacy installation script typically results in the following changes:
\enumerate{
\item \verb{~/patRoon-install}: a directory with external dependencies (\code{MetFrag}, \code{SIRIUS}, ...) and may contain \R packages from the \code{patRoonDeps} repository depending on user input during the installation.
\item \verb{~/.Rprofile-patRoon.R}: An \R script that should be loaded during \R startup, and initializes the configuration needed to use the files (1).
\item Code in \verb{~/.Rprofile} to actually load (2) on \R startup (only if it exists)
}

When \code{toggleLegacy()} is used to disable legacy installations the function simply renames the initialization script
(2) so it won't be loaded on \R startup. The \code{removeLegacy()} function removes the files from (1-2) and optionally
the changes from (3).
}

