% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{patRoonInst-package}
\alias{patRoonInst}
\alias{patRoonInst-package}
\title{Manage patRoon installations}
\description{
\Sexpr[results=text,echo=FALSE]{packageDescription("patRoonInst", fields = "Description")}
}
\section{Package options}{


The following package options (see \code{\link[=options]{options()}}) can be set:

\itemize{

\item \code{patRoonInst.repos.patRoonDeps} The URL to the \code{patRoonDeps} \R package repository.

\item \code{patRoonInst.repos.runiverse} The URL to the \href{https://r-universe.dev/}{r-universe} \R package repository.

\item \code{patRoonInst.path.legacy} The directory where legacy files were installed.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rickhelmus/patRoonInst}
  \item \url{https://rickhelmus.github.io/patRoonInst/}
  \item Report bugs at \url{https://github.com/rickhelmus/patRoon/issues}
}

}
